2021 - day 22

+++ 2021/day22/test.h
@@ -4,5 +4,5 @@
 
 namespace aoc2021_day22 {
     int part_1(std::string_view path);
-    int part_2(std::string_view path);
+    int64_t part_2(std::string_view path);
 }
